feat(bis_fetcher): add BaseFetcher and BisFetcher classes
These classes support the fetching of links and articles from websites. The BaseFetcher provides a standard way to fetch generic links and articles, while the BisFetcher extends the BaseFetcher for the specific case of the BIS. This includes methods to parse page links and article text. In addition, it includes methods to save and load links and articles, and methods for handling duplicated links.